requirements.txt
streamlit==1.42.0
pandas==2.2.0
gspread==6.0.0
google-auth==2.27.0
matplotlib==3.8.2
